[FIX] sale_management: missing variant in SO line name with quotation template
- Go to Sales > Configuration > Settings and activate "Product Configurator"
- Go to Sales > Configuration > Attributes and create an Attribute:
  * Attribute Name: Attribute X
  * Variants Creation Mode: Never (create_variant='no_variant')
  * Attribute Values:
      - Value 1
      - Value 2
- Go to Sales > Products > Products and create a Product:
  * Product Name: Product X
  * Variants: Attribute X (Value 1, Value 2)
- Go to Sales > Configuration > Quotation Templates and create a Template:
  * Quotation Template: Template X
  * Lines:
    - Product: Product X  -  Description: XXX
- Go to Sales > Orders > Quotations and create a Quotation
- Select Template X as Quotation Template
- Add Product X in Order Lines and validate Product Configurator
The Product is added to the SO line but the selected Attribute value is not displayed
in the SO line description, as it usually does.

It comes from the fact that when using a Quotation Template, if an added Product is
defined in the Template, the description from the Template (i.e. XXX) is used instead of
the default generated description.
This is not practical as it is not possible to differentiate several Product X with
different values for Attribute X, because of its Variants Creation Mode set to Never.

opw-2322827

closes odoo/odoo#57195

X-original-commit: fa1009fc128a02c2718bb0f40baa490910533d65
Signed-off-by: Anh Thao PHAM <kitan191@users.noreply.github.com>