[FIX] web_editor: do not display quality options on SVGs/GIFs
When the width and quality option for images was moved from the
optimize-dialog to the left-panel, it was made available on all image
types. However, width and quality options don't make sense on SVGs and
would break animated gifs. This commit makes it so that quality options
are only available on JPEG and PNG images.

closes odoo/odoo#57564

X-original-commit: 031648a350ee1faf6d61846d9529bab8e825846c
Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>