[FIX] sale_timesheet: visibility of the 're-invoice expenses' field on products
Before this commit:

Currently, the 'Re-Invoice Expenses' field visible when 'Can be Expensed' is
true but after the module sale_timesheet installation, it is not working.

After this commit:

The 'Re-Invoice Expenses' field visible when 'Can be Expensed' is true even
when we have a sale_timesheet module.

LINKS

PR https://github.com/odoo/odoo/pull/57259
Task-2328493

closes odoo/odoo#57856

X-original-commit: b1839f0f492fc210fb3b1d2c42701c67ba5d9ec7
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>