[FIX] website_event_questions: add wait for confirmation step
During the `test_01_tickets_questions` a tour is started as a portal
user to reserve tickets for the newly created event.
At the end of the tour, a modal is opened and once filled, the `continue`
button is used. The problem is that the tour stops without waiting but
the python tests continues. If the HttpCase session is cleared before
the tickets are registered, the test fails with a `400 Bad Request:
Session expired` warning and the next python assertion cannot be
fullfilled.

closes odoo/odoo#57400

X-original-commit: d8fcdccf4139fe5d36f8f1b0accc7366ab3576f7
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>