[FIX] base: x2many edition on res.config.settings form views
Before this, a one2many relation in a settings view would not work.
After the save, the added data in the field would be cleared.

Now, the data is still present.

With the change made to BaseSettingsModel the 'base_settings_tests' need
to only use the res.config.settings as fake data model. (relations towards
other models are still ok though). The test file is therefore altered
to comply to this new rule.

Note that this commit only validates the good behavior on the frontend side,
as a one2many relation towards a res.settings can't be done simply in
python as it is a transient model (forbidden).

Task ID: 2274168

closes odoo/odoo#57401

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>