[IMP] event_sale: speed up event price_subtotal computation
This commit improves the performance of the 'sale_price_subtotal' field of
event.event by refactoring the way it's computed.

The previous naive implementation looped on sale.order.lines linked to the
event and performed a currency conversion for every single line. This leads
up to thousands of requests to fetch the conversion rate and sum up everything
on a real production database like Odoo.com .

This new implementation groups the amounts by currency and performs a single
conversion per currency (factored by the number of events in 'self'). It has
the drawback of using 'today' to fetch conversion rates instead of the actual
SO date like previous implementation.

We (aka: all main POs as this decision was discussed) consider that as an
acceptable trade-off to increase performances.

Task ID-2335287

closes odoo/odoo#58650

X-original-commit: 9987b0f3fdaa5df8eb45ebfb9b03ceb9af820d52
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>