[IMP] web: log error when generating a report fails
For reports fetched through this endpoint, the error would only be
reported to the client, which may well ignore it entirely (e.g. show a
completely generic message which might as well be unhelpful and at
worst aggressively misleading).

Log the error locally before returning it to the client, so the info
is at least in the logs.

closes odoo/odoo#57411

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>