[IMP] auth_signup: avoid disabled users to fail an installation
When a module creates a disabled user during its installation or upgrade
(e.g. demo data), the UserError "You cannot perform this action on an
archived user" breaks the whole installation process.

We don't care if the mail cannot be sent for users created by an
installation, so let's just skip the whole function if we are installing
a module.

closes odoo/odoo#59036

X-original-commit: 4581fd3fa4faa9974c579bb26d553dbf6ec9f656
Signed-off-by: Paul Morelle <madprog@users.noreply.github.com>