[ADD] l10n_ch_qriban: QR-IBAN field on bank account
QR-IBAN can be different than IBAN, so we added a separate field
to be able to put the QR-IBAN.  It can be in handy in cases of
reconciliation afterwards or when different systems are used
at the same time.

If this field is filled in, we will use it for generating the
QR-IBAN barcode instead of the Account Number.

Thanks to Ravi Gohil.

closes odoo/odoo#57707

Task: 2307262
X-original-commit: c91714e1951fc8ef827de466dbd57b90b95f55cc
Signed-off-by: Josse Colpaert <jco@openerp.com>